Update to bootstrap 4
The GUI is going to use bootstrap 4. At the same time the GUI is
becoming a single page application so arkivar.html will now take the
role of the files arkiv.html, arkivliste.html, arkivdeliste.html and
arkivdel.html. These files will be removed when arkivar.html is
finished. This commit has the functionality of arkiv.html and
arkivliste.html.