Change how users are added on startup.
The PK creation mechanism is causing some confusion. In H2 there seems
to be no problem when you first add users in an import file and then
add users later. In mysql this is a problem as it causes an attempt to repeat an existing PK.

So I could again change the PK creation strategy or I can just programatically add users. This commit is a temporary approach to
dealing with this issue.