AdministrativeUnit and CaseFile integration
This commit sees AdministrativeUnit implemented properly. On creation of
a user a default AdministrativeUnit is created. Ideally this should have
occurred before users are created but our students require this to happen
at the same time as each student gets their own organisation.

This commit is also a partial fix on the how AdministrativeUnit needs to
be implemented for correspondancepart*. It is not possible to split this
across two commits so everything has to be in a monster-commit.