Add missing files for failed build.
Continued work on AdministrativeUnit and the need for a sequence
generator for case files per user per year.

Manually adding administrative unit to the database so it's been
removed from the sql file. This is because we are getting hastle with
primary key generation if data is added. Seems PK counter isn't updated.