Fix issue with administrativeUnit and user identification
From earlier prototyping, the codebase was not handling
administrativeUnit and userIDs properly. This meant that a casefile
could be created without being assigned to an administrativeUnit and the
user identified as the owner of the caseFile could be anybody. This
should never have been the case and this commit combined with parts of
the previous broken commit (cf80563c54890190bb717a06ee6687073cee99c6)
fixes this issue.

This also fixes previous commit that broke the build!