Fix missing Allow header
The runtest program is reporting a missing Allow header on OPTIONS.
This takes care of that issue. However this is probably not the final
solution. This is being handled by a filter at the HTTP boundary of the
application. This probably needs to be handled at the service layer to
see what the user actually can do with such a request!