Fix problem with AdministrativeUnit
runtest discovered a problem where we were not getting a unique result
when getting the next casefile sequence number. This was because we
were searching on administrativeunit.name and this field is not unique
in the table. There were 3 tuples that had the same
administrativeunit.name. The search is now on AdministrativeUnit object
along with the year.