Fix problem with broken _links [] on DocumentDescription
Undertaking a flush during the creation of a new DocumentDescription
caused a problem where _links was empty during serialisation due to a
change in the hasCode for DocumentDescription. The change came about
when version was 0 but was incremented and set to 1. The flush was
apparently not having an effect. So the version value was set to 0 when
creating a DocumentDescription and outside the scope of the service
layer when creating a DocumentDescriptionHateoasObject the value of
version was 1. This caused trouble as HateoasNoarkObject uses a hashmap
and uses the hashcode as a key.