Handle CORS for ODdata queries
CORS is not being handled properly for odata queries
We include a temporary addition so that odata searches work from
browser. The Odata code is experimental and will be moved from
/odata/arkivstruktur/*** to /hateoas-api/arkivstruktur/** in the future.
So this code should not cause a problem if it's forgotten but it should
be removed

The problem is that a request e.g. /odata/arkivstruktur/arkiv will
result in a null value for methods as the only endpoint registered
is /odata/arkivstruktur. So we "reset" servletPath to something that
is usable for OPTIONS.