Fix sequence number for fileId/mappeId
Sequence number generation for fileId/mappeId was previously designed
in a way that meant the change to a new year stopped the creation of
sequences. This was because the primary key was a single column and not
a composite key including the year. This fix spans multiple files and
includes slight change to the database structure.