Fix problem with type identification
The underlying interface mechanism is not working well here. This is
because within the domain model, the "types" of things are not as
straight forward as you require them in code. Record -> RegistryEntry,
DocumentDescription and DocumentObject are all not NoarkGeneralEntity
and will have to be handled differently, as is done here.