Fix CorrespondencePart update
Coverity showed some issues that lead me to start fixing this. But this
was really because we earlier had problems with the domain model
implementation, so a lot of code was commented out. This
became a large commit as there were lots of small changes and it was
difficult to identify a single point where it would make sense to commit
individual files. This should result in nikita being able to create and
update CorrespondencePartPerson and CorrespondencePartUnit.
CorrespondencPartInternal is still not implemented. This as because of
issues relating to AdministrativeUnit, which probably is fixed now and
it's possible to finalise CorrespondencPartInternal.