Update CorrespondencePart* code to be more in line with Single Responsibility Principle
The previous approach was a bit messy where the RegistryEntry class was
doing to much, In line with SRP it made sense to move the code over.

Two side effects that got included here.
1. Some final cleaning steps with updating of constants made their way
in here.
2. CorrespondencePartInternal received more attention.

I thought it was better to commit this in its current shape rather than
mix in too many different 'threads of thought'.