Make code base respect X-Forwarded-Host, X-Forwarded-Proto and X-Forwarded-Port
The old approach was a hard-coded public URL that nikita used. This is
the correct approach to solving the issue and places that solution as
close as possible to code that needs it.

There was some unlucky naming where the URL was called contextPath in
some places. This has been cleared up. Also the need then for a
constructor that took the contextPath also fell away.

Note:
For the yml files publicAddress now does not require contextPath
appended to it. Rather it reuses the contextPath value defined earlier
in the file. There should *not* be a slash at the end of the
publicAddress value.

Search/replace to get this functionality caused problems with the OData
antlr code. Easiest way to fix this was to remake the OData java files.