Fix update problem
We discovered that a number of update calls were setting the "new" with
the old value. This was a silly bug that has been removed. Note we see
that only the mandatory fields should have a check for null. A lot
of values that could legally be set to null can't be set to null at the
moment as there is a null check. This should be fixed in the next commit