Make Service classes extend NoarkService
This makes the design cleaner but also allows us to reduce the amount
of duplicated code in the codebase. IntelliJ Idea was reporting numerous
instances of duplicated code that, in particular, made this change
applicable.

Most importantly, this change will allow us to code the OData stuff in
NoarkService and all sub-classes will have the ability to filter using
OData. The next step will be to make the Controllers OData aware and
capable of passing the OData query to subclasses of NoarkService. This
will simplify finalising the roll out of OData support to all entities.

Note this commit also includes minor formatting changes to ensure all
classes have a similar expectancy/formatting with regards to
readability.