Role OData out to all entities
Previously OData was only possible when it was hardcoded to the various
endpoints. This was a cumbersome approach and ultimately would result in
a unnecessary mess of query parameters at endpoints and would have seen
a lot of redundant code added. So here we create a single OData endpoint
and forward any query requests to the endpoint. There is still a little
more tidying and commenting to finish up, this commit is just to get the
code in place.

For this to work we needed a filter to do the (server side) forwarding,
reflection and annotations to make things easier.