Implement DocumentDescription parent REL/HREF properly
The interface standard always had the ability to get back up to a parent
from a child. It was just misunderstood by me. This starts the
implementation process of rolling out the ability to get up to the root
of the application.

This had a number of unintended knock-on effects. Primarily I am no
longer returning entity objects outside of a transactional context. This
makes sense as we will start getting loading exceptions so entity
objects should only be returned to the controller layer if all that is
left is to serialise them and that is not the case.