Don't print out links on empty objects
This one kinda surprised me that it wasn't an issue before. This is one
of the deep classes that never change and a change here can have subtle
consequences on the implementation.  As I was testing I found null
pointer exceptions being thrown whenever a null value was set as the
entity. I'm guessing we just caught this and never allowed a null value
in here before. But in making the code more readable it makes sense
that it should be possible for a null value (the database returned 0
results) to go in and then just not add an entity to create links for.

Note: Have this change in the back of your mind if the links part of the
application start to behave funny!