Make code more sensible
There are two approaches to creating hateoas links:

 1. Identify the actual parent with systemId in the href
 2. Point to an endpoint where the parent will be identified

So for Series, the approach looks like the following:

 1. api/arkiv/1234
 2. api/arkivdel/4321/arkiv

We are choosing 1 as the approach to create more semantic rich
payloads. The client sees the exact object it needs to interact with,
not just an address to an object. This may come at a performance cost
in terms of multiple database lookups, but for now the agreed approach
is to provide a semantic rich payload.