Update object names
There is a mapping of Norwegian (OData names) to English names as used
in the domain model. Some of these were wrong and we were unable to run
OData queries using them. This fixes the problem. Note for attributes
there are still a large number that have not been translated.