Delete unused file
It will likely confuse people reading the codebase. It was there because
we had a really funny spring setup before and for some reason it was
needed. We cleaned up our spring setup, but forgot that this was still
in use. It was basically saying we were using annotations not XML for
configuration. This is picked up automatically.