Initial transformation of systemId as primary key and update model
Undertake a lot of housekeeping to move the domain model away from
integers as primary key and to use UUID/systemID instead. This is to
make it easier when undertaking OData requests and other manual queries
so that searches are done on the primary key rather than candidate key
with subsequent matching based on primary keys. It's a simplification
to make other parts of the codebase easier to code.

Hibernate has some interesting annontations @CreatedBy and
@LastModifiedBy that simplify the codebase. These were included here.