Better way to handle problems when JSON payloads are created
It should not be possible to have a null value in the database where
nullable = false in the column definition. But to be on the safe side
we tidy the implementation of writing mandatory values so that if they
are null an exception will go via out exception handler and not just
result in an anonymous null pointer in the log.

If this occurs during a GET it should not be a problem, but if it occurs
during a POST then hibernate should role back the data created.