Update for n5v5
This is another one of those difficult updates. We discovered a lot
of different problems that had to be dealt with. It started with trying
to get Part to work and there were lots of small problems along the way
that seemed interrelated so there was no good point to do a commit.