Fix FormatCode problem
FormatCode are a little bit more tricky as they are not a single string.
Rather they are two strings with a slash in the middle e.g. fmt/95

http://localhost:8092/noark5v5/hateoas-api/metadata/format/fmt/95/

This commit also show the fragility in handling the CORS stuff ourselves
as we really should not have to be updating servletpath like we are.
This will need to be revisted at some stage.