Implement ny-saksmappe as per spec
Although not properly described, it is known by us that, CaseStatus and
the other metadata entities should not use reference integrity on
metadataentity fields. Rather the list of values is presented to the
client under _links and when POSTing a sakssatus value is copied.

Potentially we do one thing wrong, and that is to copy in the kodenavn
value, if it is unknown.