Introduce support for OIDC description
The Noark standard does not really explain the use of OIDC
so we are basing this code on online examples. Some of this is
correct but we don't know if this is a complete minimal
example of what a OIDC compliant Noark core would support.

Note. There is no refrech token endpoint defined in OIDC,
rather a POST to oauth/token with the following portion of
a grant should be in place:
 ... grant_type=refresh_token& ...
We don't support a unserinfo endpoint either.

Remember that Basic Auth is covering the oauth/token endpoint.
This "secret" is not exposed here so clients will need to
obtain this via another channel.

Includes a few other fixes:

1. Discovered that we are still using nikita rel rather arkivverket rel. Changed.
2. Move OAuthController to a more logical place in code (with the other controllers).