Tidy up building of rel/href
We discovered during the upgrade to n5v5 from n5v4 that there
were multiple definitions of the base strings that define rels
and href in nikita. This just made things messy when trying to
upgrade and will make it messy next time. This fixes the
problem by removing duplicate definitions.