Remove version as comparable field
There was a problem having version as a comparable field as once the database transaction boundary was encountered. The version that was used in a hashmap did not generate the same checksum after the transaction was completed and the payload was to be written as json. This fixes the problem.