Fix user account creation problem
When creating a user account we discovered we were missing a
anonymousUser principal. This is required in the case where no user has logged on to create an account. So if anonymousUser is detected, then the account creation is assumed to be done by creation. It is likely this should be
revisited. Perhaps after the account is created, system can be swapped out with the actual username.