Fix OData problems
OData has been sketchy for a while now as we tried to do a major lift of the OData code and were left with something that didn't work anymore, even for basic stuff. This is the first attempt to get the code working again.

We will need a new round now where we examine this code in detail and write a description of the approach