Add new CI job to clone noark5tester and execude runtest.
Run nikita in the background, and wait for it to listen for
HTTP requests before running runtest.  Running the test in
a Debian image to get the required software available.

Use UTF-8 locales when executing runtest.  This is
not enough to get print() statements to handle unicode
strings, but will ensure python know which charset to use.

Allow this new test job to fail until we fix all the issues
it reports.