Fix New Year change in sequence number and record number
When a new year starts nikita  will automatically create a new sequnce number and record number for the administrative unit when they are required. This is not left as an administrative task by the record keeper sp that nikita can continuously operate uninterrupted, even over the change to a new year.

The change from 2019 to 2020 showed that this was not implemented properly. The call to getNextCaseFileSequence on the newly generated sequence generator caused a "a different object with the same identifier value was already associated ..." error. This I believe was because two sequence generators were being created. Initially we suspected a problem in the domain model, but the fix is to make sure that two sequence generators are not being persisted. A later commit should include an actual test to prove that this is not a potential problem in production.