Handle author inline in serialization and deserialisation.
There seem to be two conflicting approaches for updating author
information.  One using separate endpoints announced in the hateoas
handlers, and one inline in the deserialiser (but not the
serialisers).  The deserialisers do not currently seem to work, but
assuming inline is the wanted appraoch, this patch
make sure to serialise author information and remove the author related
relation keys from _links and the code.