Implement support for ElectronicSignature.
Dropped relation links to related unwanted endpoints and activate
parse testing for all relevant entities.

Tried to fix search too, but only for the metadata code name, not
for the code value.  Not sure how to fix that, and suspect this should
be fixed the same way for all metadata.

Corrected some field name capitalisation in the process.