Add forgotten relations to metadata/ _links list.
The relations to access restriction and disposal decision are used
elsewhere in the API, but not listed in the metadata list. The end
points are not yet implemented.

Not implementing the end points, as the metadata machinery need
major changes, and it seemed less useful to add a new controller
based on the current implementations.