Remove codebeat  from pipeline
Currently codebeat seems to take a long time to finish >8 hours and sometimes does not finish. We are not really looking at the codebeat results at the moment anyway so we are removing it for now. I hope to bring it in again if we can get it to run faster and finish.