Make download more robust.
Before this change, GET on the download href before doing any POST would
cause nikita to return 500 error because of an exception.

After this change, the download code will first check if there is a
file to return, then decide which HTTP headers to set to reflect
the available file.  This allow the code to handle better if no
mime type, file name or size is available.

Hopefully fixes #139.