Remove wrong print method
printNikitaEntity should never have been here. It was confusing that entities could have one of two possibilities for printing. This caused a knock on effect to force other serialisers to use the correct method. Also removed NoarkGeneralEntitySerializer.java as it is not in use. It was left over from early implementation.