Implement series precursor/successor handling.
Allow precursor and successor reference strings to refer to
series not currently in the archive by simply storing the
provided SystemID as UUIDs, and only add relations in _links
to the precursor and successor if the series exist in the archive.

Move related constants from Constants.java to N5ResourceMappings.java
alongside the other Series field names.