Change way metadata entities are handled at service and controller
There is so much boilerplate code on the metadata entities, and they are all just/copy paste of the same basic code with some constants changed. This makes updating things a pain as we end up grep/replace all the time. This commit is the first step in reducing the controller and service classes down to one class for all metadata entities.

Note there is now a temporary fake endpoint for testing purposes metadata_X. This will be removed. It's there because I can't get the existing controller to work with the new one in parallel.

The change to pom.xml is that there is a version dependency to the Reflection class somewhere in spring I believe and including th dependency fixes a problem with null pointer deep in code.

Fix PUT/update and GET for ny-*

Drop temporary fix in URL namespace to respond to