Finalise _links approach
Code is working so wrapping this up. Added comments to explain
approach and throw a 500 if nikita hits a problem when trying
to serialise an object here. Typically the problem that can be
encountered is that we are either missing a annotation.

Note the approach to getting the request address is the one that
needs to be used in nikita and avoid hardcoding a value in
application*.yml files.