<Temp change hardcoded value for dokumenttype
We are experiencing a problem when creating a documentdescription
from the GUI.
Object [id=B] was not of the specified subclass [nikita.common.model.noark5.v5.metadata.DocumentType]
 : loaded object was of wrong class class nikita.common.model.noark5.v5.metadata.DocumentStatus; nested exception is org.hibernate.WrongClassException: Object
 [id=B] was not of the specified subclass [nikita.common.model.noark5.v5.metadata.DocumentType] : loaded object was of wrong class class nikita.common.model.noark5.v5.metadata.DocumentStatus

This is likely a result of spring not being able to distinguish between metadata repositories in the same transaction. I am unsure why it doesn't distinguish within the same repository. I just need a simple example for teaching and replacing B with R makes the problem go away. B is used in both documentstatus aswell as documenttype. The problem occurs because the fetch of "B" is returning "B" in documentstatus.

This will be fixed, just making sure the problem is clearly logged here. >