Initial commit of updated approach to OData
This introduces testing of OData that allows us to approach OData
development using a TDD approach. Currently tests are limited to
generating and verifying expected HQL syntax, but later we expect
to test against data in a database as well and perhaps to use the
hibernate CriteriaBuilder. I am not comfortable using the CriteriaBuilder
until I get a wider understanding of what has to be supported.