Add more OData functionality
This commit includes the ability to do JOINs, better support for integer search.
Reflection is introduced in the Noark domain model to aid with JOINS. We have not
dealt with JOINs for ManyToMany and the code is still a bit rough. See the comment close to the start of ODataParser.g4 for information about the next steps.