Create test for known metadata problem
It is known that there is a problem when persisting two metadata
entities during a single session. The problem manifests itself
when two similar codes are in use, but the problem is there all
the time.

This commit is to make a test so that we can study the code and start
the process of fixing the problem.

See:
https://gitlab.com/OsloMet-ABI/nikita-noark5-core/-/issues/184