Make HateoasNoarkObject serializable
Normally this class would not be serializable, but when writing an
empty OData query, we do not know what type of object to instantiate.
There is a need to ensure that this object is serialised correctly
so that the proper format is written. This change forces the correct
code to be executed.