Abandon approach for deprecated find method
findValidMetadataByEntityTypeOrThrow(entityType, code, codename) was earlier marked as deprecated. The reason to deprecate the method was to make the code easier to read by passing in the readily available IMetadataEntity object rather then pulling out the code and codename pairs. So another approach was introduced, findValidMetadataByEntityTypeOrThrow(entityType, IMetadataEntity metadataEntity). After reading through the code there is only one case where the code is using the 'wrong' approach. This fixes that. There is nothing wrong with either approach, but for readability it makes sense to only have one entry point.

See also:
https://gitlab.com/OsloMet-ABI/nikita-noark5-core/-/issues/190