Fix OpenAPI documentation and tidy up a bit
Swagger/OpenAPI support has been a problem in nikita for a long time. I was never able to get the springfox swagger setup to work. It just hangs on startup, preventing nikita from starting. Therefore swagger was left uninitialized but the annotations were left in the code. A while back I saw a new library was launched and was was waiting for time to test it. This updates OpenAPI support in nikita making OpenAPI descriptions available from nikita without logging in. If running from localhost see:

   http://localhost:8092/noark5v5/swagger-ui
   http://localhost:8092/noark5v5/v3/api-docs

It takes 3-4 minutes to generate the openAPI response, so we likely need to see about generating the documentation offline and making that available.

Note: If you are looking through this commit, you will see formatting and other changes that are not really to do with swagger. I know it's bad practice to mix these kind of things, but for a long time I have been hoping to tidy up the code in the @Controller classes and trying to make them look similar to a reader coming into the codebase. The OpenAPI upgrade required me to go through each controller and if I had not done the others thing now, I would likely not find the time to revisit these classes and tidy up. The tidy up also includes removing a lot of warnings so it was worth it. Also no functionality changes happened, the changes are more cosmetic and to make sure we actually can show the API to an explorer.

Note also that OpenAPI is exposing too much information from the domain model that is not required to understand how nikita works. Now that we have gotten OpenAPI to work, perhaps we can look at how the documentation works. Remember nikita also generates documentation when you run test with spring-restdocs. Even though the documentation there is limited, it may still provide interesting information.

So this commit fixes OpenAPI support and tidies things up a bit for readability