Fix more bidirectional synchronization issues
There are a lot of 'CRITICAL - BidirectionalSynchronizationEvent' events and these are being dealt with in order.

Note: This is the last commit with these fixes
Note 2: These commits just introduce the ability to use the add / remove approach with bidirectional associations being handled correctly. It's likely that the code has done this 'manually'. So there should be a new round of checks where we use add* of object rather than set* of Lists.