Tidy up in use of User
This is an interesting commit to be aware of as it does something not really expected in the tidying up of the domain model. The User class had @ManyToMany(fetch = FetchType.EAGER) for user->authority. The eager was there because we there was an issue with a lazy initialisation exception being thrown a long time ago. SO and other sources provided a solution to make the relationship EAGER and the problem was solved. At the time, due to time constraints, I accepted that explanation. Now when running the optimiser we get a warning that this is a bad choice so I remeoved it and the lazy exception came up again. A new round of studying the problem leads me to believe that one of the classes seems to be out of the transaction scope. We could probably label it a @Component and make sure the transaction is there, but I had to ask why we had a NikitaPrincipal. Why are we using something additional to what spring provides, but not doing anything useful with it. So removing it and going back to the spring user means results in no exception and things  are running much more smoothly. This change is not really that big as it has no impact other than likely increased performance, but it's worth flagging among the various similar changes that are occurring.