Fix Comment ManyToMany issue
Comment like many other classes should have a different approach to ManyToMany. This resulted in a number of knock changes. Especially make note of CommentHateoasHandler update and the use of OData query to get Comment related entities.

This also brings in two new OData tests. The eq test shows the approach this commit takes should work, while the other is left as documentation that there is a problem doing UUID like queries that needs to be adressed.