Fix OData endpoint exposure
nikita uses an internal forwarding mechanism to handle all OData queries.
It was the simplest way to solve the problem of having an endpoint for OData
queries. An earlier commit replaced request.getServletPath() with UrlPathHelper().getPathWithinApplication(request) as there are tests that had trouble resolving
the servlet path during testing. The change resulted in the hidden odata address
being exposed back to clients when running runtest. runtest does not test OData
queries at the moment so this wasn't discovered until now. The new approach is to
try and use getServletPath(), and if that fails use the UrlPathHelper. Later we will
likely revisit the way OData is dealt with and try to have a base class on the controllers
that can handle the OData queries.