Fix ManyToMany performance issue for User / AdministrativeUnit / Authority
The optimiser reports a 'CRITICAL - ManyToManyListEvent' for the ManyToMany use of User / AdministrativeUnit / Authority. Using a Set instead of a List may result in increased performance.

This commit also deals with some of the intricacies of hibernate and spring.
The optimiser advised to make sure both ends of manytomany were being set.
We did this and started seeing problems when adding a user. My
understanding is that the
<p>
     admin.addAdministrativeUnit(administrativeUnit);
<p>
which looks like:
<p>
      public void addAdministrativeUnit(AdministrativeUnit  administrativeUnit) {
        this.administrativeUnits.add(administrativeUnit);
        administrativeUnit.getUsers().add(this);
      }
<p>
as causing a problem when we did
<p>
       if (userExists(user.getUsername())) {
           throw new UsernameExistsException("There is an account with that " +
           "username: " + user.getUsername());
       }
<p>
as the username is available in the current session and a get query will
persist the data. The solution applied here then is to simply avoid the the
get here as the save will result in a DataIntegrityViolationException
exception being thrown if the user already exists in the database.