Standardise OneToOne approach
This has been a commit with a lot of offline work. I believe the entities
PostalAddress <-> CorrespondencePartUnit and CorrespondencePartPerson are
in a polymorphic association. This means the @MapsId approach is not possible
for PostalAddress as it is not possible to take the same FK from both CorrespondencePartPerson and CorrespondencePartUnit down PostalAddress as a PK. It could be possible as both
inherit from CorrespondencePart, but I do not have to explore this issue any more.

This commit also includes some house-keeping and tidying up.