Update BSM tests
The latest work and update making the tests transactional have meant the original approach uses various values. This tides up in this. Also I want to start using the WithCustomMockUser approach for test going forward so this brings the necessary code in place.

Note, there is something strange about our setup. The SecurityContext created is being ignored in the tests so for now I have a hack of updating the username of the anonymousUser for the request. Yes, this is a bad approach, but time constraints mean I have to wrap this up. We can revisit this later.