Fix return type issue
A test containing `fondsRepository.delete(fonds);` failed and made me realise that the temporary 'hack' we had for dealing with returning systemID was not OK for spring-data. When we changed the type for systemId from String to UUID it broke a lot code so we wrapped the getSystemId as a string. This means that when spring use reflection and calls entity.getSystemId()  it would get the wrong type (String, but should be UUID).