Ensure compliance with API standard
The object was not in compliance. Added missing values and tidied up code.
Also putting systemId here as I am testing why adding en electronic
signature to a DocumentDescription is failing. Need to look at something else
for a little why, but that is the reason why this change with systemId is here.