Move name from child classes to parent class
The inheritance was not working. A search for part/navn only returned
one of child object (partPerson OR partUnit). It was partPerson, perhaps
alphabetical order is the reason.  On thinking about it, name being in the
two children means they are two independent fields that just share the same
name. Moving name to part means that a search on part/navn will search across
both.

This also results in updated tests and default database values.