Replaced 'oppdatert*' with 'endret*' everywhere to match latest spec
The extraction specification (Noark 5) added 'endretDato' and 'endretAv'
to most entities, and the API specification replaced all use of
'oppdatertDato' and 'oppdatertAv' to follow suit, so Nikita should be
changed too.

See https://github.com/arkivverket/noark5-standard/issues/55 and
https://github.com/arkivverket/noark5-tjenestegrensesnitt-standard/issues/255
for the details.  The API changes are currently incomplete, but expected to
be completely migrated soon.