Drop join requirement support null search
I wasn't aware that we need to support both types of BSM search

   * api/arkivstruktur/virksomhetsspesifikkeMetadata?$filter=ppt-v1:skolekontakt ne null
   * api/arkivstruktur/mappe?$filter=virksomhetsspesifikkeMetadata/ppt-v1:skolekontakt ne null&$top=1

The latter is a JOIN, while the first is a normal entity search.

The isNull change is to be able to pick up the isNullValue attribute for the search of null values. 'Value' gets added later so this approach uses the same string'Value', boolean'Value'