Temp allow test to fail
Test is failing as 200 is being returned instead of 400. This is because the way BSM search works is that it occurs as keys in the database at runtime and it's acceptable for a nul value to be returned. We need to make the test fail if a non-existent attribute is used.