Fix OData search problem for variantformat
The internal name translator was failing an OData query
because variantformat was being returned instead of
variantFormat.

Note: We need to see if it is possible to merge the
DOCUMENT_OBJECT_VARIANT_FORMAT and VARIANT_FORMAT
constants. One might be used for metadata, the other
in documentObject, so there might be a need to keep
them seperate.