Add ability to override version
When doing an expand to casefile, we had a problem with version not being set and
throwing version exception. This allows us to not experience that problem. However,
this method should not really be used elsewhere in code. It is required to handle
a particular case e.g. creating a dummy casefile to return