Update GUI for simulated email import
For teaching we need to be able to show the following process:

1. Nikita receives email
2. Email is converted to a CaseFile and sent to casehandler
3. Casehandler finalises case handling and sends case to leader
4. Leader approves
5. Case goes to Records manager for final approval

This is a large fix, with redundant code that was added to get the solution
to work for teaching. We simulate email, but the approach can be combined
the the import-email tool or pere's auto forwarding of email to nikita.

The code needs to be improved on significantly! Also the html should really be combined
into a single page app.