Fix proxy problem when looking for object class names
We discovered that hibernate/spring proxying of classes caused
problems when trying to find the class name to create a join.
An example is Series was being reported as

    Series$HibernateProxy$xsdrs.

and the code could no join to the series tables. This does not allow
us to re-use the name properly. This is also left here as a hint that
perhaps we should re-evaualte the reflections approach. Perhaps we
could use annotations in some way.