Auto convert all records to registryentry when doing an expand
For the example of teaching it makes sense when expanding a
File to CaseFile, you need the records to become registryEntry.
It is not clear that this is the correct approach and perhaps
the client should themselves decide what is to be a RegistryEntry.
Keeping the approach in master for the time being as there are good
arguments for and against the approach.

The approach however raises the need for other types of conversion.

* Record -> RecordNote
* RecordNote -> RegistryEntry
* RegistryEntry -> RecordNote
* RegistryEntry -> Record

RegistryEntry -> Record would loose data.