Fix concurrency issue
We are seeing a concurrency issue where setVersion is flagging
and throw an exception because the ETAG is not in sync. I believe
we saw something similar in the CaseFile expand-to-casefile work.
We need to dig a bit deeper to solve this, but this commit gets rid
of the issue.

https://gitlab.com/OsloMet-ABI/nikita-noark5-core/-/issues/195