Change autocommit approach
The optimiser is informing that the JDBC Connection was borrowed in auto-commit mode, and Hibernate needs to disable this mode when starting a transaction, and re-enable it prior to releasing the JDBC Connection back to the pool.

This is noted as MINOR issue, but also an easy one to fix.

See the following for more details:
https://vladmihalcea.com/why-you-should-always-use-hibernate-connection-provider_disables_autocommit-for-resource-local-jpa-transactions/