Change @Transactional approach
Avoid marking the entire class as @Transactional and only mark methods that really are @Transactional as @Transactional. The optimiser informs that the JDBC Connection was borrowed for [1] ms, but no SQL statement was executed. This is a problem as it deprives other concurrent requests of using the connection.

This means that any new method added to this class must know whether or not it is a simple read and does not change data or actually requires a transaction.