Introduce CaseFileExpansion
When undertaking a utvid-til-saksmappe, the previous approach was based on using creating JSON at the service layer. This was not a good approach, but worked. We discovered that we were missing a _links and then it would be very messy to start working on URLs here. A much cleaner way to solve the problem is taken here.