Update approach
After earlier changing the @Transactional approach, a problem with duplicate entries (multiple users with same email) was causing exceptions. This, I believe, was not a problem before as the transactional boundary and the catch took care of it. With the updated @Transactional approach the adding of users became a problem again.

The if (userService.userExists()) ... approach is the correct approach to use here and it is what should have been in place from earlier.

Note: There is a difference when persisting a new User from here (where we reuse an existing AdministrativeUnit) and when adding a new user via the GUI where a new AdministrativeUnit is being declared at the same time.