Make approach work from one class
The prior approach with the actual insert being a seperate class stems from issues surrounding duplicates being inserted in the database. This was the symptom and not the cause as the real problem likely was the same as we experienced previously with it not being possible to update two metadata entities in the same transaction. That update means that MetadataInsert works as expected.